+++ packages/pros/src/devices/adi/digital.rs
@@ -5,6 +5,54 @@ use pros_sys::PROS_ERR;
 use super::{AdiDevice, AdiDeviceType, AdiError, AdiPort};
 use crate::error::bail_on;
 
+/// Represents the logic level of a digital pin.
+///
+/// On digital devices, logic levels represent the two possible voltage signals that define
+/// the state of a pin. This value is either [`High`] or [`Low`], depending on the intended
+/// state of the device.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum LogicLevel {
+    /// A high digital signal.
+    ///
+    /// ADI ports operate on 3.3V logic, so this value indicates a voltage of 3.3V or above.
+    High,
+
+    /// The low digital signal.
+    ///
+    /// ADI ports operate on 3.3V logic, so this value indicates a voltage below 3.3V.
+    Low,
+}
+
+impl LogicLevel {
+    /// Returns `true` if the level is [`High`].
+    pub const fn is_high(&self) -> bool {
+        match self {
+            Self::High => true,
+            Self::Low => false,
+        }
+    }
+
+    /// Returns `true` if the level is [`Low`].
+    pub const fn is_low(&self) -> bool {
+        match self {
+            Self::High => false,
+            Self::Low => true,
+        }
+    }
+}
+
+impl core::ops::Not for LogicLevel {
+    type Output = Self;
+
+    fn not(self) -> Self::Output {
+        match self {
+            Self::Low => Self::High,
+            Self::High => Self::Low,
+        }
+    }
+}
+
+/// Generic digital input ADI device.
 #[derive(Debug, Eq, PartialEq)]
 /// Generic digital input ADI device.
 pub struct AdiDigitalIn {
@@ -17,30 +65,26 @@ impl AdiDigitalIn {
         Self { port }
     }
 
-    /// Gets a rising-edge case for a digital button press.
-    pub fn new_press(&mut self) -> Result<bool, AdiError> {
-        Ok(unsafe {
-            bail_on!(
-                PROS_ERR,
-                pros_sys::ext_adi_digital_get_new_press(
-                    self.port.internal_expander_index(),
-                    self.port.index()
-                )
-            ) != 0
+    /// Gets the current logic level of a digital input pin.
+    pub fn level(&self) -> Result<LogicLevel, AdiError> {
+        let value = bail_on!(PROS_ERR, unsafe {
+            pros_sys::ext_adi_digital_read(self.port.internal_expander_index(), self.port.index())
+        }) != 0;
+
+        Ok(match value {
+            true => LogicLevel::High,
+            false => LogicLevel::Low,
         })
     }
 
-    /// Gets the current value of a digital input pin.
-    pub fn value(&self) -> Result<bool, AdiError> {
-        Ok(unsafe {
-            bail_on!(
-                PROS_ERR,
-                pros_sys::ext_adi_digital_read(
-                    self.port.internal_expander_index(),
-                    self.port.index()
-                )
-            ) != 0
-        })
+    /// Returns `true` if the digital input's logic level level is [`LogicLevel::High`].
+    pub fn is_high(&self) -> Result<bool, AdiError> {
+        Ok(self.level()?.is_high())
+    }
+
+    /// Returns `true` if the digital input's logic level level is [`LogicLevel::Low`].
+    pub fn is_low(&self) -> Result<bool, AdiError> {
+        Ok(self.level()?.is_high())
     }
 }
 
@@ -60,8 +104,8 @@ impl AdiDevice for AdiDigitalIn {
     }
 }
 
-#[derive(Debug, Eq, PartialEq)]
 /// Generic digital output ADI device.
+#[derive(Debug, Eq, PartialEq)]
 pub struct AdiDigitalOut {
     port: AdiPort,
 }
@@ -72,15 +116,29 @@ impl AdiDigitalOut {
         Self { port }
     }
 
-    /// Sets the digital value (1 or 0) of a pin.
-    pub fn set_value(&mut self, value: bool) -> Result<i32, AdiError> {
-        Ok(bail_on!(PROS_ERR, unsafe {
+    /// Sets the digital logic level (high or low) of a pin.
+    pub fn set_level(&mut self, level: LogicLevel) -> Result<(), AdiError> {
+        bail_on!(PROS_ERR, unsafe {
             pros_sys::ext_adi_digital_write(
                 self.port.internal_expander_index(),
                 self.port.index(),
-                value,
+                level.is_high(),
             )
-        }))
+        });
+
+        Ok(())
+    }
+
+    /// Set the digital logic level to [`LogicLevel::High`]. Analagous to
+    /// [`Self::set_level(LogicLevel::High)`].
+    pub fn set_high(&mut self) -> Result<(), AdiError> {
+        self.set_level(LogicLevel::High)
+    }
+
+    /// Set the digital logic level to [`LogicLevel::Low`]. Analagous to
+    /// [`Self::set_level(LogicLevel::Low)`].
+    pub fn set_low(&mut self) -> Result<(), AdiError> {
+        self.set_level(LogicLevel::Low)
     }
 }